+++ Modules/Intent.Modules.Aws.Sqs/Templates/SqsEventBus/SqsEventBusTemplatePartial.cs
@@ -36,7 +36,7 @@ namespace Intent.Modules.Aws.Sqs.Templates.SqsEventBus
                 .AddUsing("Microsoft.Extensions.Options")
                 .AddClass($"SqsEventBus", @class =>
                 {
-                    @class.ImplementsInterface(this.GetEventBusInterfaceName());
+                    @class.ImplementsInterface(this.GetMessageBusInterfaceName());
                     @class.AddField("IAmazonSQS", "_sqsClient", field => field.PrivateReadOnly());
                     @class.AddField("List<MessageEntry>", "_messageQueue", field => field.PrivateReadOnly().WithAssignment(new CSharpStatement("[]")));
                     @class.AddField("Dictionary<string, PublisherEntry>", "_lookup", field => field.PrivateReadOnly());
@@ -60,14 +60,34 @@ namespace Intent.Modules.Aws.Sqs.Templates.SqsEventBus
                         method.AddStatement("_messageQueue.Add(new MessageEntry(message));");
                     });
                     
+                    @class.AddMethod("void", "Publish", method =>
+                    {
+                        method.AddGenericParameter("T", out var T);
+                        method.AddGenericTypeConstraint(T, c => c.AddType("class"));
+                        method.AddParameter(T, "message");
+                        method.AddParameter("IDictionary<string, object>", "additionalData");
+
+                        method.AddStatement("// Note: AWS SQS does not support additional data in this implementation, ignoring parameter");
+                        method.AddStatement("Publish(message);");
+                    });
+                    
                     @class.AddMethod("void", "Send", method =>
                     {
                         method.AddGenericParameter("T", out var T);
                         method.AddGenericTypeConstraint(T, c => c.AddType("class"));
                         method.AddParameter(T, "message");
 
-                        method.AddStatement("ValidateMessage(message);");
-                        method.AddStatement("_messageQueue.Add(new MessageEntry(message));");
+                        method.AddStatement("Publish(message);");
+                    });
+                    
+                    @class.AddMethod("void", "Send", method =>
+                    {
+                        method.AddGenericParameter("T", out var T);
+                        method.AddGenericTypeConstraint(T, c => c.AddType("class"));
+                        method.AddParameter(T, "message");
+                        method.AddParameter("IDictionary<string, object>", "additionalData");
+
+                        method.AddStatement("Publish(message, additionalData);");
                     });
 
                     @class.AddMethod("Task", "FlushAllAsync", method =>